feat(ui): sidebar shell
The app's collapsible sidebar (provider, icon-collapsed mode, tooltip
menu buttons, mobile sheet, width constants) moves to @stll/ui/sidebar;
the app renders it with its own copy and hotkey wired in.
