+++ .changeset/ui-sidebar-shell.md
@@ -1,0 +1,5 @@
+---
+"@stll/ui": minor
+---
+
+Add a generic, host-agnostic sidebar shell at `@stll/ui/sidebar`: `SidebarProvider`/`useSidebar` (controlled or uncontrolled open state, mobile sheet, icon-collapse), `Sidebar`, `SidebarHeader`/`Content`/`Footer`/`Separator`, the `SidebarGroup*` family, the `SidebarMenu*` family (with a tooltip shown only while collapsed), `SidebarInset`, `SidebarRail`, `SidebarTrigger`, and the `SIDEBAR_WIDTH_PX`/`SIDEBAR_WIDTH_ICON_PX` size constants.
