+++ .changeset/workspace-frame-sidebar-doc.md
@@ -1,0 +1,5 @@
+---
+"@stll/workspace-ui": patch
+---
+
+Document that `WorkspaceFrame`'s `host-responsive` composition is where a host mounts an expandable/collapsible sidebar (such as the shell now published at `@stll/ui/sidebar`), while `described` always renders the narrower, fixed-width `ApplicationRail`.
