+++ apps/web/package.json
@@ -115,7 +115,6 @@
     "@vscode/observables": "0.1.1-0",
     "better-auth": "catalog:",
     "better-result": "catalog:",
-    "class-variance-authority": "^0.7.1",
     "cobe": "^2.0.1",
     "diff": "catalog:",
     "driver.js": "1.8.0",
