+++ apps/web/src/components/sidebar-sizing.ts
@@ -1,2 +1,1 @@
-export const SIDEBAR_WIDTH_PX = 256;
-export const SIDEBAR_WIDTH_ICON_PX = 48;
+export { SIDEBAR_WIDTH_ICON_PX, SIDEBAR_WIDTH_PX } from "@stll/ui/sidebar";
