+++ oxlint.config.ts
@@ -2324,7 +2324,6 @@ export default defineConfig({
         "apps/web/src/**/kanban-column.tsx",
         "apps/web/src/**/workspace-table.tsx",
         "apps/web/src/**/workspace-table/**/*.tsx",
-        "apps/web/src/**/sidebar.tsx",
         "apps/web/src/**/template-preview.tsx",
         "apps/web/src/**/page-citation.tsx",
         // Generated message types: UI copy may legitimately contain words
