+++ packages/ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@stll/ui",
-  "version": "0.19.0",
+  "version": "0.18.0",
   "description": "Stella's design system: bidi-aware React primitives built on Base UI, the dockable inspector pane, and the Tailwind v4 theme they are styled with.",
   "keywords": [
     "base-ui",
@@ -86,6 +86,7 @@
     "./select": "./src/components/select.tsx",
     "./separator": "./src/components/separator.tsx",
     "./sheet": "./src/components/sheet.tsx",
+    "./sidebar": "./src/components/sidebar.tsx",
     "./loader": "./src/components/loader.tsx",
     "./skeleton": "./src/components/skeleton.tsx",
     "./stella-mark": "./src/components/stella-mark.tsx",
@@ -146,6 +147,7 @@
     "./components/select": "./src/components/select.tsx",
     "./components/separator": "./src/components/separator.tsx",
     "./components/sheet": "./src/components/sheet.tsx",
+    "./components/sidebar": "./src/components/sidebar.tsx",
     "./components/skeleton": "./src/components/skeleton.tsx",
     "./components/stella-mark": "./src/components/stella-mark.tsx",
     "./components/stella-wordmark": "./src/components/stella-wordmark.tsx",
